web/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
torch>=2.0.0
torchaudio>=2.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
transformers>=4.30.0
numpy>=1.24.0
Pillow>=10.0.0

